Move test_pyro_camera into test_huntsman.py, remove test_observatory.py
